pip_requirements.txt
django== 1.3.1#1.6.2
django-pagination
psycopg2
django-notification==0.2
PIL
gdata==2.0.15